+++ test/hexapdf/font/true_type/table/test_hmtx.rb
@@ -5,7 +5,7 @@ require_relative 'common'
 require 'hexapdf/font/true_type/table/hhea'
 require 'hexapdf/font/true_type/table/hmtx'
 
-describe HexaPDF::Font::TrueType::Table::Hhea do
+describe HexaPDF::Font::TrueType::Table::Hmtx do
   before do
     data = [1, -2, 3, -4, 5, -6].pack('ns>ns>s>2')
     set_up_stub_true_type_font(data)
@@ -17,14 +17,14 @@ describe HexaPDF::Font::TrueType::Table::Hhea do
   describe "initialize" do
     it "reads the data from the associated file" do
       table = create_table(:Hmtx)
-      assert_equal(1, table[0].advance_width)
-      assert_equal(-2, table[0].left_side_bearing)
-      assert_equal(3, table[1].advance_width)
-      assert_equal(-4, table[1].left_side_bearing)
       assert_equal(3, table[2].advance_width)
       assert_equal(5, table[2].left_side_bearing)
       assert_equal(3, table[3].advance_width)
       assert_equal(-6, table[3].left_side_bearing)
+      assert_equal(1, table[0].advance_width)
+      assert_equal(-2, table[0].left_side_bearing)
+      assert_equal(3, table[1].advance_width)
+      assert_equal(-4, table[1].left_side_bearing)
     end
   end
 end